+++ CHANGES.md
@@ -9,6 +9,7 @@
 
 * update FastAPI requirement to `>=0.87`
 * update rio-tiler requirement to `>=4.0,<4.1`
+* remove `rescale` and `color_formula` from the `post_process` dependency
 
 **breaking changes**